requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
click==7.1.2
flask-serialize==1.5.0
flask-sqlalchemy==2.4.4
flask==1.1.2
itsdangerous==1.1.0
jinja2==2.11.3
markupsafe==1.1.1
permissive-dict==1.0.3
psycopg2==2.8.6
sqlalchemy==1.3.23
werkzeug==1.0.1
python-dotenv
requests
authlib
six
python-jose
flask-cors
